.exrc: use ^A as a prefix key
I do not like ^X because it involves my pinky. Pressing ^X with ring
and and index fingers is not always comfortable. Sometimes it is, but
it depends on my left palm's position.

---

^A: conveys the meaning of the function more clearly.

---

Which is better: ^Axss or ^Asxx?

Typically, by mistake, people run privileged commands from unprivileged
users. Thus, they have to correct themselves.

^Axx -> ^Asxx almost looks like 'cmd' -> 'sudo cmd'.

[s]^Asxe is more awkward to type than ^Axse. (But ^Asx} is not.)[/s]
Actually, no. If you press 'sx' in quick succession, hitting 'e'
is natural.

On the other hand, ^Ax emphasizes the particular mode in which the
command is run.

^Axx -> ^Axss in both cases sounds like "print the command's output".

For a moment, let's try ^As.