Use one environment variable to identify if a SOCKS server is running.
I do not need multiple environment variables. With .chatgptrc.d/proxy,
I need to configure a proxy so that requests to openai.com work. With
.apirc.include.d/mirai-proxy, I need to configure a proxy so that requests
to miraitech.co resources work. With .mos.sh.include.d/mirai-proxy, I
need to configure a proxy so that ssh connections to miraitech.co work. In
each case, it is the user who cares about the details, and not the tool.