.exrc: get rid of #:
The purpose of these functions is to use them in combination with other
macros, e.g. define a region and run a command on it. This way, it is
possible to define just one macro doing a thing, rather many instances
operating on different ranges. (With my shell macros, this actually
makes sense. With per-project macros, used sparingly, it's clutter.)

`#:` just does not fit this purpose. For regular functions, consider
the default.