.exrc: implement shell shortcuts as marked region functions
# 'v!'w vs 'v!`w

There is a difference. With the former, the filter includes the line
on which mark w is set. This is consistent with the principle of marked
regions, thus preferred. With the latter, the command includes everything
until mark w.

This behavior is mentioned by William Joy, albeit indirectly: [1].

Consider the implications of both alternatives.

When I do #e, I look for the whole string, for correctness. !'w includes
the string, !`w does not. When I pass #e to shell, the output either
follows the EOF (!'w) or comes before it (!`w).

When I do #}, mark w points to an empty line. !'w includes it, !`w does
not. When I pass #} to shell, the output either follows the empty line
(!'w) or comes before it (!`w).

It should also be noted that it is impossible to have an empty line
after EOF unless it's already there.

I could've left #}#A and #}#S as is, for ease of configuration and,
possibly, quick jumps to output. But it is not always consistent in
itself. The newline wouldn't appear if } jumps to the end of file and
it is not an empty line.

In general, #A and #S should use the !'w form to support all possible
regions. I assume that the full format like #}#A would be used
occasionally, at least, not as endlessly as shell. Implementing shell
shortcuts with marked region functions seems more like a proof-of-concept.

And it does not seem possible to either make #{ exclude new lines or
make #e include the found pattern.

1: An Introduction to Display Editing with Vi, 8.4