.exrc: use unused characters for building macros
If I take @a@, typing the normal @a produces a one second delay.

I think it is impossible to find a combination that's universally
free. _su_ breaks someone else's _s. To this end, vim functions is the
right tool. What is possible here is to find a combination that's unused
by both vi/vim and this config.

In vim, \ is actually used as <Leader>. However, it does not seem to be
actually mapped to anything, which is also stated in `:help index`.

Also rework ChatGPT shortcuts. Use #![a-zA-Z] space for global shortcuts.